Create Average_Weekly_Temperature
This VB project is used to calculate the average weekly temperatures. Users enter each week's temperature into text boxes and select the "calculate" button for the average to be calculated. The "Clear" button is used to reset the text boxes and average calculation to the default settings and "Exit" button to terminate the application.