Create TG_Automotive2
This application is an extension of the TG Automotive project. This breaks up the application into two forms where users first selection from a list of preferences which are summarized in the bottom of the form. The second half is providing more preferences for the user to select from and finally calculating the cost break downs.