Create Student_Clubs2
This application is an extension from the original Student Clubs project. This allows a user to add student names to a list of various club membership options providing a real-time calculation of total members of each group. The program saves each student member name so users are free to move in and out of various groups member lists.