Create WEb_App
This project is a design for users to view charges for a hotel reservation. The application displays the complete breakdown of charges to the customer. The application allows a user to enter information such as additional fees, how many nights a customer stays and what the nightly charge is to calculate the breakdown. This VB project was later transitioned into a web forum (ASP.NET) using HTML code to build a basic web application.